requirements.txt
quart
aiohttp
PyPDF2
reportlab
google-cloud-secret-manager
google-cloud-firestore
uvicorn
python-dateutil
cryptography
